Split a UnitCommandWidget from WorldCommandWidget
