+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -559,6 +559,7 @@
     <Compile Include="UtilityCommands\ExtractSettingsDocsCommand.cs" />
     <Compile Include="UtilityCommands\ExtractTraitDocsCommand.cs" />
     <Compile Include="Widgets\Logic\Ingame\MusicControllerLogic.cs" />
+    <Compile Include="Widgets\UnitCommandWidget.cs" />
     <Compile Include="WorldExtensions.cs" />
     <Compile Include="UtilityCommands\GetMapHashCommand.cs" />
     <Compile Include="UtilityCommands\Glob.cs" />
